@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -9,14 +9,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="IdentityServer4" Version="4.1.2" />
-    <PackageReference Include="Rsk.Saml" Version="5.0.0" />
-    <PackageReference Include="Rsk.Saml.IdentityServer4" Version="5.0.0" />
+    <PackageReference Include="Duende.IdentityServer" Version="6.3.8" />
+    <!--<PackageReference Include="Rsk.Saml" Version="6.2.2" />-->
+    <!--<PackageReference Include="Rsk.Saml.IdentityProvider" Version="6.2.2" />-->
+    <PackageReference Include="Rsk.Saml.DuendeIdentityServer" Version="6.4.1" />
 
-    <PackageReference Include="Serilog" Version="2.9.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
-    <PackageReference Include="Serilog.Sinks.File" Version="4.1.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="6.1.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
+    <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
   </ItemGroup>
 
   <ItemGroup>
