@@ -13,6 +13,7 @@ namespace idpWithIdpInitiated
 {
     public class Program
     {
+        [Obsolete]
         public static void Main(string[] args)
         {
             Console.Title = "IdentityServer4";
@@ -20,6 +21,7 @@ namespace idpWithIdpInitiated
             BuildWebHost(args).Run();
         }
 
+        [Obsolete]
         public static IWebHost BuildWebHost(string[] args)
         {
             return WebHost.CreateDefaultBuilder(args)
@@ -27,10 +29,10 @@ namespace idpWithIdpInitiated
                     .UseSerilog((context, configuration) =>
                     {
                         configuration
-                            .MinimumLevel.Debug()
-                            .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                            .MinimumLevel.Override("System", LogEventLevel.Warning)
-                            .MinimumLevel.Override("Microsoft.AspNetCore.Authentication", LogEventLevel.Information)
+                            .MinimumLevel.Verbose()
+                            //.MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
+                            //.MinimumLevel.Override("System", LogEventLevel.Warning)
+                            //.MinimumLevel.Override("Microsoft.AspNetCore.Authentication", LogEventLevel.Information)
                             .Enrich.FromLogContext()
                             .WriteTo.File(@"identityserver4_log.txt")
                             .WriteTo.Console(outputTemplate: "[{Timestamp:HH:mm:ss} {Level}] {SourceContext}{NewLine}{Message:lj}{NewLine}{Exception}{NewLine}", theme: AnsiConsoleTheme.Literate);
