@@ -1,4 +1,4 @@
-﻿@using IdentityServer4.Extensions
+﻿@using Duende.IdentityServer.Extensions
 @{
     string name = null;
     if (!true.Equals(ViewData["signed-out"]))
