@@ -1,2 +1,2 @@
-﻿@using IdentityServer4.Quickstart.UI
+﻿@using Duende.IdentityServer.Quickstart.UI
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
